Simple helper function
a function named index_range that takes two integer arguments page and page_size.
The function returns a tuple of size two containing a start index and an end index corresponding to the range of indexes to return in a list for those particular pagination parameters. Page numbers are 1-indexed, i.e. the first page is page 1.