Simple pagination
a method named get_page that takes two integer arguments page with default value 1 and page_size with default value 10.
Uses assert to verify that both arguments are integers greater than 0.
Uses index_range to find the correct indexes to paginate the dataset correctly and return the appropriate page of the dataset (i.e. the correct list of rows).